Remove unneeded interface
